refactor(adapters): split services into adapters + services
Create qfa.adapters package and move the LiteLLM client into it.
qfa.services now contains only application services (orchestrator
and the prompt-building helpers in coding_classifier); driven adapter
implementations live in qfa.adapters.

This change is a pure rename; behaviour is unchanged.

- new qfa.adapters package with a docstring describing the role
- git mv src/qfa/services/llm_client.py -> src/qfa/adapters/llm_client.py
- git mv tests/services/test_llm_client.py -> tests/adapters/test_llm_client.py
- update test patches qfa.services.llm_client.* -> qfa.adapters.llm_client.*
- update qfa.api.app to import LiteLLMClient from qfa.adapters
